services/rag-service/requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
groq>=0.4.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
